Split large app entrypoints into focused modules
