+++ council/app/tui/app.py
@@ -3,24 +3,16 @@ from __future__ import annotations
 import asyncio
 import uuid
 from dataclasses import replace
-from pathlib import Path
 from typing import Literal
 
 from textual import on
-from textual.app import App, ComposeResult
+from textual.app import App
 from textual.binding import Binding
-from textual.containers import Vertical
-from textual.message import Message
-from textual.screen import ModalScreen
-from textual.widgets import Footer, Header, RichLog, Static, TextArea
+from textual.widgets import Footer, Header, RichLog
 
 from council.adapters.bus.in_memory import InMemoryBus
-from council.adapters.clock.system import SystemClock
-from council.adapters.packs.filesystem import FilesystemPackLoader
-from council.adapters.storage.json_file import JsonFileStore
-from council.app.bootstrap import build_elders
-from council.app.config import load_config
 from council.app.tui.council_view import CouncilView
+from council.app.tui.widgets import CouncilInput, SynthesizerModal
 from council.domain.debate_service import DebateService
 from council.domain.draft_analysis import analyze_drafts
 from council.domain.events import (
@@ -38,46 +30,6 @@ from council.domain.ports import (
     ElderPort,
     TranscriptStore,
 )
-
-
-class CouncilInput(TextArea):
-    """A TextArea where Enter submits and Ctrl+Enter inserts a newline."""
-
-    class Submitted(Message):
-        def __init__(self, value: str) -> None:
-            super().__init__()
-            self.value = value
-
-    BINDINGS = [
-        # priority=True so our Enter handler runs BEFORE TextArea's built-in
-        # enter-inserts-newline binding.
-        Binding("enter", "submit", "Submit", show=False, priority=True),
-        Binding("ctrl+enter", "newline", "Insert newline", show=False),
-    ]
-
-    def action_submit(self) -> None:
-        self.post_message(self.Submitted(self.text))
-
-    def action_newline(self) -> None:
-        self.insert("\n")
-
-
-class SynthesizerModal(ModalScreen[ElderId]):
-    BINDINGS = [
-        Binding("1", "pick('ada')", "Ada"),
-        Binding("2", "pick('kai')", "Kai"),
-        Binding("3", "pick('mei')", "Mei"),
-        Binding("escape", "dismiss(None)", "Cancel"),
-    ]
-
-    def compose(self) -> ComposeResult:
-        yield Vertical(
-            Static("Who should synthesize?"),
-            Static("[1] Ada   [2] Kai   [3] Mei   [Esc] Cancel"),
-        )
-
-    def action_pick(self, elder: str) -> None:
-        self.dismiss(elder)  # type: ignore[arg-type]
 
 
 class CouncilApp(App):
@@ -100,6 +52,7 @@ class CouncilApp(App):
         Binding("3", "focus_pane('mei')", "Mei", show=False),
         Binding("4", "focus_pane('synthesis')", "Synthesis", show=False),
     ]
+    _ELDER_LABELS: dict[ElderId, str] = {"ada": "Ada", "kai": "Kai", "mei": "Mei"}
 
     def __init__(
         self,
@@ -120,11 +73,6 @@ class CouncilApp(App):
         self._pack_loader = pack_loader
         self._pack_name = pack_name
         self._using_openrouter = using_openrouter
-        # "r1_only" (default): stop after R1, await user decision. The 2026-04
-        # experiments found this is the strongest shape for diverse rosters;
-        # it's also the drafting-friendly mode (three independent drafts, no
-        # flattening). "full" runs R1+R2 back-to-back as the legacy TUI did,
-        # then awaits user decision.
         self._mode: Literal["r1_only", "full"] = mode
         self._prev_cost_total: float = 0.0
         self._bus = InMemoryBus()
@@ -132,7 +80,7 @@ class CouncilApp(App):
         self._debate: Debate | None = None
         self.awaiting_decision: bool = False
         self.is_finished: bool = False
-        self.rendered_lines: list[str] = []  # test-observable notice buffer
+        self.rendered_lines: list[str] = []
         self._tasks: set[asyncio.Task] = set()
         self._view = CouncilView(clock=clock)
         self._report_store = report_store
@@ -143,7 +91,7 @@ class CouncilApp(App):
         task.add_done_callback(self._tasks.discard)
         return task
 
-    def compose(self) -> ComposeResult:
+    def compose(self):
         yield Header()
         yield RichLog(id="notices", markup=True, wrap=True, highlight=False)
         yield self._view
@@ -152,7 +100,7 @@ class CouncilApp(App):
 
     async def on_mount(self) -> None:
         self._stream_task = self._spawn(self._consume_events())
-        self.query_one("#input", CouncilInput).focus()
+        self._input().focus()
         self._spawn(self._run_health_checks())
 
     async def on_unmount(self) -> None:
@@ -161,56 +109,63 @@ class CouncilApp(App):
         for task in list(self._tasks):
             task.cancel()
 
-    # --- event fan-out ---------------------------------------------------
     async def _consume_events(self) -> None:
         async for ev in self._bus.subscribe():
             if isinstance(ev, TurnStarted):
-                self._view.pane(ev.elder).begin_thinking(ev.round_number)
+                self._on_turn_started(ev)
             elif isinstance(ev, TurnCompleted):
-                self._view.pane(ev.elder).end_thinking_completed(ev.answer, questions=ev.questions)
-                # Fan each outgoing question into the TARGET elder's pane.
-                for q in ev.questions:
-                    self._view.pane(q.to_elder).on_incoming_question(q)
+                self._on_turn_completed(ev)
             elif isinstance(ev, TurnFailed):
-                self._view.pane(ev.elder).end_thinking_failed(ev.error)
+                self._on_turn_failed(ev)
             elif isinstance(ev, RoundCompleted):
-                # r1_only mode: decision point after R1 (no auto-R2).
-                # full mode: decision point after R2 (R1+R2 auto-chain).
-                decision_at = 1 if self._mode == "r1_only" else 2
-                if ev.round.number >= decision_at:
-                    self.awaiting_decision = True
-                    self.query_one("#input", CouncilInput).disabled = False
-                    # First time the user sees decision affordances, surface
-                    # the option summary. Keeps discoverability without
-                    # cluttering every round.
-                    if ev.round.number == decision_at:
-                        self._write_decision_hint()
-                    # Auto-synth modal when all three elders converge (R3+).
-                    if ev.round.number >= 3 and self._service.rules.is_converged(ev.round):
-                        self.run_worker(self._synthesize_worker(), exclusive=True)
-                if self._using_openrouter:
-                    self._spawn(self._write_cost_notice())
+                self._on_round_completed(ev)
             elif isinstance(ev, SynthesisCompleted):
-                self._view.pane("synthesis").end_thinking_completed(ev.answer)
-                self._view.pane("synthesis").focus()
-                self.is_finished = True
-                self.awaiting_decision = False
-                # Auto-generate a debate report and write it to disk.
-                if ev.answer.elder:
-                    self._spawn(self._generate_and_write_report(ev.answer.elder))
+                self._on_synthesis_completed(ev)
             elif isinstance(ev, UserMessageReceived):
-                # Dispatch to all three elder panes for inline rendering.
-                for pane_key in ("ada", "kai", "mei"):
-                    self._view.pane(pane_key).on_user_message(ev.message)
+                self._on_user_message_received(ev)
 
-    # --- health check ----------------------------------------------------
+    def _input(self) -> CouncilInput:
+        return self.query_one("#input", CouncilInput)
+
+    def _on_turn_started(self, ev: TurnStarted) -> None:
+        self._view.pane(ev.elder).begin_thinking(ev.round_number)
+
+    def _on_turn_completed(self, ev: TurnCompleted) -> None:
+        self._view.pane(ev.elder).end_thinking_completed(ev.answer, questions=ev.questions)
+        for q in ev.questions:
+            self._view.pane(q.to_elder).on_incoming_question(q)
+
+    def _on_turn_failed(self, ev: TurnFailed) -> None:
+        self._view.pane(ev.elder).end_thinking_failed(ev.error)
+
+    def _on_round_completed(self, ev: RoundCompleted) -> None:
+        decision_at = 1 if self._mode == "r1_only" else 2
+        if ev.round.number >= decision_at:
+            self._set_awaiting_decision(True)
+            if ev.round.number == decision_at:
+                self._write_decision_hint()
+            if ev.round.number >= 3 and self._service.rules.is_converged(ev.round):
+                self.run_worker(self._synthesize_worker(), exclusive=True)
+        if self._using_openrouter:
+            self._spawn(self._write_cost_notice())
+
+    def _on_synthesis_completed(self, ev: SynthesisCompleted) -> None:
+        self._view.pane("synthesis").end_thinking_completed(ev.answer)
+        self._view.pane("synthesis").focus()
+        self.is_finished = True
+        self._set_awaiting_decision(False)
+        if ev.answer.elder:
+            self._spawn(self._generate_and_write_report(ev.answer.elder))
+
+    def _on_user_message_received(self, ev: UserMessageReceived) -> None:
+        for pane_key in ("ada", "kai", "mei"):
+            self._view.pane(pane_key).on_user_message(ev.message)
+
+    def _set_awaiting_decision(self, value: bool) -> None:
+        self.awaiting_decision = value
+        self._input().disabled = not value
+
     async def _run_health_checks(self) -> None:
-        labels: dict[ElderId, str] = {
-            "ada": "Ada",
-            "kai": "Kai",
-            "mei": "Mei",
-        }
-
         async def _probe(elder_id: ElderId) -> tuple[ElderId, bool]:
             try:
                 ok = await self._elders[elder_id].health_check()
@@ -224,11 +179,11 @@ class CouncilApp(App):
             return
         for eid in unhealthy:
             self._write_notice(
-                f"[yellow]⚠ {labels[eid]} CLI is unavailable or unauthenticated. "
+                f"[yellow]⚠ {self._ELDER_LABELS[eid]} CLI is unavailable or unauthenticated. "
                 f"Install it and run its `login` command before asking a question.[/yellow]"
             )
         if len(unhealthy) == len(self._elders):
-            self.query_one("#input", CouncilInput).disabled = True
+            self._input().disabled = True
             self._write_notice(
                 "[red]No elders available. Fix the vendor CLI setup above, "
                 "then restart the app.[/red]"
@@ -239,10 +194,6 @@ class CouncilApp(App):
         self.query_one("#notices", RichLog).write(line)
 
     def _write_decision_hint(self) -> None:
-        """Surface the keybinding affordances once per debate, at the
-        first decision point. Non-intrusive but discoverable — the
-        keybindings themselves are ``show=False`` in the Footer.
-        """
         if self._mode == "r1_only":
             self._write_notice(
                 "[dim]R1 complete. [d] compare drafts (agreements/divergences)  ·  "
@@ -284,16 +235,14 @@ class CouncilApp(App):
         )
         self._write_notice(f"[blue]{line}[/blue]")
 
-    # --- user actions ----------------------------------------------------
     @on(CouncilInput.Submitted)
     async def _on_input_submitted(self, event: CouncilInput.Submitted) -> None:
         text = event.value.strip()
         if not text:
             return
-        input_widget = self.query_one("#input", CouncilInput)
+        input_widget = self._input()
         input_widget.clear()
         if self._debate is None:
-            # First submission — use as the initial prompt.
             pack = self._pack_loader.load(self._pack_name)
             self._debate = Debate(
                 id=str(uuid.uuid4()),
@@ -307,13 +256,11 @@ class CouncilApp(App):
             self._view.focus()
             self._spawn(self._opening_exchange())
         else:
-            # Between-round user message.
             if not self.awaiting_decision:
                 return
             self._spawn(self._service.add_user_message(self._debate, text))
 
     async def _generate_and_write_report(self, by: ElderId) -> None:
-        """Generate the debate report and render/save it."""
         if self._debate is None:
             return
         try:
@@ -330,30 +277,18 @@ class CouncilApp(App):
                 self._write_notice(f"[yellow]Report file write failed: {ex}[/yellow]")
 
     async def _opening_exchange(self) -> None:
-        """Run the initial rounds dictated by ``self._mode``.
-
-        r1_only: run R1 only, then hand control back to the user. This is
-        the default — the 2026-04 experiments found R1-only-then-maybe-
-        synthesise is the strongest shape for diverse rosters, and it's
-        drafting-friendly (three independent takes, no flattening).
-
-        full: run R1 then R2 back-to-back (the legacy TUI behavior). The
-        RoundCompleted handler keeps input disabled between R1 and R2
-        because decision_at is 2 in this mode.
-        """
         if self._debate is None:
             return
-        await self._service.run_round(self._debate)  # R1
+        await self._service.run_round(self._debate)
         if self._debate.status != "in_progress":
             return
         if self._mode == "full":
-            await self._service.run_round(self._debate)  # R2
+            await self._service.run_round(self._debate)
 
     async def action_continue_round(self) -> None:
         if not self.awaiting_decision or self._debate is None:
             return
-        self.awaiting_decision = False
-        self.query_one("#input", CouncilInput).disabled = True
+        self._set_awaiting_decision(False)
         self._spawn(self._service.run_round(self._debate))
 
     async def action_abandon(self) -> None:
@@ -361,7 +296,7 @@ class CouncilApp(App):
             return
         self._debate.status = "abandoned"
         self.is_finished = True
-        self.awaiting_decision = False
+        self._set_awaiting_decision(False)
         self.exit()
 
     async def action_override(self) -> None:
@@ -376,13 +311,6 @@ class CouncilApp(App):
         self.run_worker(self._synthesize_worker(), exclusive=True)
 
     async def action_analyze_drafts(self) -> None:
-        """Compare R1 drafts — show agreements/divergences/unique points.
-
-        Available after R1 completes. Uses Ada as the analyzer (the same
-        default as the synthesiser rotation). Output is appended to the
-        synthesis pane; synthesis can still be run afterwards and will
-        appear below the analysis.
-        """
         if not self.awaiting_decision or self._debate is None or not self._debate.rounds:
             return
         self.run_worker(self._analyze_drafts_worker(), exclusive=False)
@@ -407,12 +335,7 @@ class CouncilApp(App):
         choice = await self.push_screen_wait(SynthesizerModal())
         if choice is None:
             return
-        self.awaiting_decision = False
-        # Disable the input so keystrokes don't get swallowed by the TextArea
-        # while the user is waiting for synthesis.
-        self.query_one("#input", CouncilInput).disabled = True
-        # Reveal the synthesis pane (columns: display=True; tabs: dynamically
-        # adds the tab) and focus it so the ticker is immediately visible.
+        self._set_awaiting_decision(False)
         await self._view.show_synthesis_pane()
         self._view.pane("synthesis").begin_thinking(round_number=1)
         self._view.pane("synthesis").focus()
@@ -429,64 +352,9 @@ class CouncilApp(App):
 
 
 def main() -> None:
-    import argparse
-    import os
+    from council.app.tui.cli import main as cli_main
 
-    parser = argparse.ArgumentParser(prog="council")
-    parser.add_argument("--pack", default="bare")
-    parser.add_argument("--packs-root", default=str(Path.home() / ".council" / "packs"))
-    parser.add_argument("--store-root", default=str(Path.home() / ".council" / "debates"))
-    parser.add_argument("--reports-root", default=str(Path.home() / ".council" / "reports"))
-    parser.add_argument(
-        "--claude-model",
-        default=os.environ.get("COUNCIL_CLAUDE_MODEL"),
-        help="Model alias or full name passed to `claude --model` (e.g. sonnet, opus).",
-    )
-    parser.add_argument(
-        "--gemini-model",
-        default=os.environ.get("COUNCIL_GEMINI_MODEL"),
-        help="Model name passed to `gemini -m` (e.g. gemini-2.5-flash — recommended; Pro has tight quota).",
-    )
-    parser.add_argument(
-        "--codex-model",
-        default=os.environ.get("COUNCIL_CODEX_MODEL"),
-        help="Model name passed to `codex exec -m` (e.g. gpt-5-codex).",
-    )
-    parser.add_argument(
-        "--mode",
-        choices=["r1_only", "full"],
-        default="r1_only",
-        help=(
-            "Debate mode. 'r1_only' (default) runs R1 then stops for you "
-            "to read the three drafts, synthesise with `s`, or continue "
-            "to a cross-examination round with `c`. 'full' runs R1+R2 "
-            "back-to-back as the legacy TUI did."
-        ),
-    )
-    args = parser.parse_args()
+    cli_main()
 
-    packs_root = Path(args.packs_root)
-    packs_root.mkdir(parents=True, exist_ok=True)
-    (packs_root / args.pack).mkdir(exist_ok=True)  # ensure bare pack works
 
-    config = load_config()
-    cli_models: dict[ElderId, str | None] = {
-        "ada": args.claude_model,
-        "kai": args.gemini_model,
-        "mei": args.codex_model,
-    }
-    elders, using_openrouter, _roster_spec = build_elders(config, cli_models=cli_models)
-
-    from council.adapters.storage.report_file import ReportFileStore
-
-    app = CouncilApp(
-        elders=elders,
-        store=JsonFileStore(root=Path(args.store_root)),
-        clock=SystemClock(),
-        pack_loader=FilesystemPackLoader(root=packs_root),
-        pack_name=args.pack,
-        using_openrouter=using_openrouter,
-        report_store=ReportFileStore(root=Path(args.reports_root)),
-        mode=args.mode,
-    )
-    app.run()
+__all__ = ["CouncilApp", "CouncilInput", "main"]
